refac(rust): merge servers and loadouts files
